try1

+++ exercises/23_conversions/using_as.rs
@@ -1,11 +1,6 @@
-// Type casting in Rust is done via the usage of the `as` operator.
-// Note that the `as` operator is not only used when type casting. It also helps
-// with renaming imports.
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    // TODO: Make a conversion before dividing.
-    total / values.len()
+    total / values.len() as f64  // Correct type conversion
 }
 
 fn main() {
